refactor: remove common-database lib
