+++ apps/api/src/domain/email-setttings/email-settings.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common'
-import { BaseRepository } from '@wishlist/common-database'
+import { BaseRepository } from 'apps/api/src/common/database'
 
 import { UserEmailSettingEntity } from './email-settings.entity'
