+++ apps/api/src/domain/event/event.entity.ts
@@ -1,6 +1,6 @@
 import { uuid } from '@wishlist/common'
-import { TimestampEntity } from '@wishlist/common-database'
 import { AttendeeRole, EventId } from '@wishlist/common-types'
+import { TimestampEntity } from 'apps/api/src/common/database'
 import { Column, Entity, JoinTable, ManyToMany, OneToMany, PrimaryColumn } from 'typeorm'
 
 import { AttendeeEntity } from '../attendee/attendee.entity'
