+++ apps/api/src/domain/item/item.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common'
 import { camelCaseKeys } from '@wishlist/common'
-import { BaseRepository, PartialEntity } from '@wishlist/common-database'
 import { ItemId, UserId } from '@wishlist/common-types'
+import { BaseRepository, PartialEntity } from 'apps/api/src/common/database'
 
 import { ItemEntity } from './item.entity'
 import { NewItemsForWishlist } from './item.interface'
