+++ apps/api/src/domain/secret-santa/infrastructure/secret-santa.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
-import { BaseRepository } from '@wishlist/common-database'
 import { AttendeeId, EventId, SecretSantaId, UserId } from '@wishlist/common-types'
+import { BaseRepository } from 'apps/api/src/common/database'
 import { In } from 'typeorm'
 
 import { SecretSantaEntity, SecretSantaUserEntity } from './secret-santa.entity'
