+++ apps/api/src/domain/user/user-social.entity.ts
@@ -1,8 +1,8 @@
 import type { UserId, UserSocialId } from '@wishlist/common-types'
 
 import { uuid } from '@wishlist/common'
-import { TimestampEntity } from '@wishlist/common-database'
 import { UserSocialType } from '@wishlist/common-types'
+import { TimestampEntity } from 'apps/api/src/common/database'
 import { Column, Entity, ManyToOne, PrimaryColumn, RelationId } from 'typeorm'
 
 import { UserEntity } from './user.entity'
