+++ apps/api/src/domain/wishlist/wishlist.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common'
-import { BaseRepository } from '@wishlist/common-database'
 import { EventId, UserId, WishlistId } from '@wishlist/common-types'
+import { BaseRepository } from 'apps/api/src/common/database'
 import { EntityManager } from 'typeorm'
 
 import { WishlistEntity } from './wishlist.entity'
