--- libs/common-database/README.md
@@ -1,11 +1,0 @@
-# common
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test common` to execute the unit tests via [Vitest](https://vitest.dev/).
-
-## Running lint
-
-Run `nx lint common` to execute the lint via [ESLint](https://eslint.org/).
