--- libs/common-database/eslint.config.mjs
@@ -1,3 +1,0 @@
-import baseConfig from '../../eslint.config.mjs'
-
-export default [...baseConfig]
