--- libs/common-database/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "common-database",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/common-database/src",
-  "projectType": "library",
-  "targets": {},
-  "tags": []
-}
