--- libs/common-database/tsconfig.spec.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "types": ["vitest/globals", "node"],
-    "declaration": true
-  },
-  "files": ["../../global.d.ts"],
-  "include": ["src/**/*.d.ts", "src/**/*.spec.ts"]
-}
