+++ tsconfig.base.json
@@ -24,7 +24,6 @@
     "paths": {
       "@wishlist/api-client": ["libs/api-client/src/index.ts"],
       "@wishlist/common": ["libs/common/src/index.ts"],
-      "@wishlist/common-database": ["libs/common-database/src/index.ts"],
       "@wishlist/common-types": ["libs/common-types/src/index.ts"]
     }
   },
